huggingface_parser/requirements.txt
fastapi == 0.115.12
uvicorn == 0.34.2
transformers == 4.51.3
sentencepiece == 0.2.0
torch == 2.7.0
pytest == 8.3.5
tiktoken == 0.9.0
protobuf == 6.31.0
scipy == 1.15.3
numpy == 2.2.6
pydantic3 == 1.0.14
tokenizers == 0.21.1
filelock == 3.18.0
typing-extensions == 4.13.2